now with abstract class
