+++ src/Models/Teams2Rors.php
@@ -12,12 +12,8 @@ declare(strict_types=1);
 
 namespace Elabftw\Models;
 
-use Elabftw\Elabftw\Db;
-use Elabftw\Enums\Action;
-use Elabftw\Exceptions\IllegalActionException;
 use Elabftw\Exceptions\ImproperActionException;
 use Elabftw\Interfaces\QueryParamsInterface;
-use Elabftw\Services\Check;
 use Override;
 use PDO;
 
@@ -26,33 +22,20 @@ use function sprintf;
 /**
  * Handle the teams to ror relationship. Submodel for teams.
  */
-final class Teams2Rors extends AbstractRest
+final class Teams2Rors extends Abstract2Rors
 {
     public function __construct(
         private readonly int $teamid,
-        private readonly bool $canwrite = false,
-        private readonly ?string $ror = null,
+        bool $canwrite = false,
+        ?string $ror = null,
     ) {
-        $this->Db = Db::getConnection();
-        if ($this->ror !== null) {
-            Check::ror($this->ror);
-        }
+        parent::__construct($canwrite, $ror);
     }
 
     #[Override]
     public function getApiPath(): string
     {
         return sprintf('api/v2/teams/%d/rors/', $this->teamid);
-    }
-
-    #[Override]
-    public function postAction(Action $action, array $reqBody): int
-    {
-        $this->canWriteOrExplode();
-        return match ($action) {
-            Action::Create => $this->create(),
-            default => throw new ImproperActionException('Incorrect action for ROR.'),
-        };
     }
 
     #[Override]
@@ -88,24 +71,8 @@ final class Teams2Rors extends AbstractRest
         return $this->Db->execute($req);
     }
 
-    private function selectAll(int $teamid): array
-    {
-        $sql = 'SELECT * FROM teams2rors WHERE teams_id = :teams_id ORDER BY created_at ASC';
-        $req = $this->Db->prepare($sql);
-        $req->bindValue(':teams_id', $teamid, PDO::PARAM_INT);
-        $this->Db->execute($req);
-        return $req->fetchAll();
-
-    }
-
-    private function canwriteOrExplode(): void
-    {
-        if (!$this->canwrite) {
-            throw new IllegalActionException();
-        }
-    }
-
-    private function create(): int
+    #[Override]
+    protected function create(): int
     {
         if ($this->ror === null) {
             throw new ImproperActionException('Missing ROR value in URL');
@@ -117,5 +84,15 @@ final class Teams2Rors extends AbstractRest
         $this->Db->execute($req);
 
         return 1;
+    }
+
+    private function selectAll(int $teamid): array
+    {
+        $sql = 'SELECT * FROM teams2rors WHERE teams_id = :teams_id ORDER BY created_at ASC';
+        $req = $this->Db->prepare($sql);
+        $req->bindValue(':teams_id', $teamid, PDO::PARAM_INT);
+        $this->Db->execute($req);
+        return $req->fetchAll();
+
     }
 }
