+++ src/Models/Users2Rors.php
@@ -12,12 +12,8 @@ declare(strict_types=1);
 
 namespace Elabftw\Models;
 
-use Elabftw\Elabftw\Db;
-use Elabftw\Enums\Action;
-use Elabftw\Exceptions\IllegalActionException;
 use Elabftw\Exceptions\ImproperActionException;
 use Elabftw\Interfaces\QueryParamsInterface;
-use Elabftw\Services\Check;
 use Override;
 use PDO;
 
@@ -26,33 +22,20 @@ use function sprintf;
 /**
  * Handle the users to ror relationship. Submodel for users.
  */
-final class Users2Rors extends AbstractRest
+final class Users2Rors extends Abstract2Rors
 {
     public function __construct(
         private readonly int $userid,
-        private readonly bool $canwrite = false,
-        private readonly ?string $ror = null,
+        bool $canwrite = false,
+        ?string $ror = null,
     ) {
-        $this->Db = Db::getConnection();
-        if ($this->ror !== null) {
-            Check::ror($this->ror);
-        }
+        parent::__construct($canwrite, $ror);
     }
 
     #[Override]
     public function getApiPath(): string
     {
         return sprintf('api/v2/users/%d/rors/', $this->userid);
-    }
-
-    #[Override]
-    public function postAction(Action $action, array $reqBody): int
-    {
-        $this->canWriteOrExplode();
-        return match ($action) {
-            Action::Create => $this->create(),
-            default => throw new ImproperActionException('Incorrect action for ROR.'),
-        };
     }
 
     #[Override]
@@ -87,14 +70,8 @@ final class Users2Rors extends AbstractRest
         return $this->Db->execute($req);
     }
 
-    private function canWriteOrExplode(): void
-    {
-        if (!$this->canwrite) {
-            throw new IllegalActionException();
-        }
-    }
-
-    private function create(): int
+    #[Override]
+    protected function create(): int
     {
         if ($this->ror === null) {
             throw new ImproperActionException('Missing ROR value in URL');
